requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
ffmpeg-python==0.2.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
aiofiles==23.2.1
httpx==0.25.2
